make truecolor work in VT history viewer (fix #1388)

+++ far2l/src/console/AnsiEsc.hpp
@@ -47,7 +47,7 @@ namespace AnsiEsc
 	private:
 		Parser _parser;
 		FontState _font_state;
-		WORD _initial_attr;
+		DWORD64 _initial_attr;
 		wchar_t _last_char = L' ';
 
 		void EnforceStateColor();
