+++ far2l/src/console/interf.hpp
@@ -101,8 +101,8 @@ void SetScreen(int X1,int Y1,int X2,int Y2,wchar_t Ch,int Color);
 void MakeShadow(int X1,int Y1,int X2,int Y2);
 void ChangeBlockColor(int X1,int Y1,int X2,int Y2,int Color);
 void SetColor(int Color, bool ApplyToConsole = false);
-void SetRealColor(WORD wAttributes, bool ApplyToConsole = false);
-WORD GetRealColor();
+void SetRealColor(DWORD64 wAttributes, bool ApplyToConsole = false);
+DWORD64 GetRealColor();
 void ClearScreen(int Color);
 int GetColor();
